Restructure project for Maven
Files have been reorganized, checkstyle tests have been fixed, and failing tests have been corrected.
