+++ .gitignore
@@ -21,3 +21,9 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+#Visual Studio Code
+.vscode/
+
+#Maven
+target/
