--- code/model/Frequency.java
@@ -1,12 +1,0 @@
-package code.model;
-
-/** The Rate of which the habit can be updated
- * 
- * @author	CS 3212
- * @version Spring 2022
- */
-public enum Frequency {
-	DAILY,
-	WEEKLY,
-	MONTHLY
-}
