--- code/model/SudokuPuzzle.java
@@ -1,5 +1,0 @@
-package code.model;
-
-public class SudokuPuzzle {
-    
-}
