+++ code/src/main/java/habit_mode/model/Frequency.java
@@ -1,0 +1,12 @@
+package habit_mode.model;
+
+/** The Rate of which the habit can be updated
+ * 
+ * @author	CS 3212
+ * @version Spring 2022
+ */
+public enum Frequency {
+	DAILY,
+	WEEKLY,
+	MONTHLY
+}
