+++ code/src/main/java/habit_mode/model/SudokuPuzzle.java
@@ -1,0 +1,5 @@
+package habit_mode.model;
+
+public class SudokuPuzzle {
+    
+}
