+++ code/src/main/java/habit_mode/view/codebehind/LoginScreen.java
@@ -1,0 +1,40 @@
+package habit_mode.view.codebehind;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.TextField;
+
+public class LoginScreen {
+
+    @FXML
+    private ResourceBundle resources;
+
+    @FXML
+    private URL location;
+
+    @FXML
+    private Button loginButton;
+
+    @FXML
+    private TextField passwordTextField;
+
+    @FXML
+    private TextField userNameTextField;
+
+    @FXML
+    void loginButtonPress(ActionEvent event) {
+
+    }
+
+    @FXML
+    void initialize() {
+        assert this.loginButton != null : "fx:id=\"loginButton\" was not injected: check your FXML file 'LoginScreen.fxml'.";
+        assert this.passwordTextField != null : "fx:id=\"passwordTextField\" was not injected: check your FXML file 'LoginScreen.fxml'.";
+        assert this.userNameTextField != null : "fx:id=\"userNameTextField\" was not injected: check your FXML file 'LoginScreen.fxml'.";
+
+    }
+
+}
